Streaming merge try 1
This approach works correctly if we have a single row of tiles. Then it
can merge them in a streaming approach.

It only merges two tiles at once and they have to be adjacent. For example
in Morton-order at the 3rd tile in a quad, the 3. is merged with the 2.
but not merged with the 1. because it is not loaded into memory.

Also, only check the local max. point ID, but not the global
max. point ID, which is problematic when merging the 1+3 tiles
for example

The memory profile (profiling/mprofile_20191210184938.dat) looks allright,
but need to test with larger data so I can see a longer running process.