Streaming merge try 2
Solved the problem with including all of the adjacent tiles as the base (a8adcfc20183ab9c7389c0888af45f1fcb125474).
Merging works, process runs through, memory profile looks alright.

There is a problem that extra triangles are created in the stars of the points that are the
densified vertices along the border, when the .triangles() method is called. For example
37fz_2 and star 1 shouldn't have triangle (1,3,1730) because that's not in the obj (no 'f 2 4 1731').

It might be because of this, but there are also many many None vertices in the links of
the stars after I start merging the TINs and putting the base-es together into a single Star.